Rename HTTPS implementation to H2
