+++ crates/resolver/src/lib.rs
@@ -259,11 +259,11 @@ pub mod config;
 pub mod dns_lru;
 pub mod dns_sd;
 pub mod error;
+#[cfg(feature = "dns-over-https")]
+mod h2;
 #[cfg(feature = "dns-over-h3")]
 mod h3;
 mod hosts;
-#[cfg(feature = "dns-over-https")]
-mod https;
 pub mod lookup;
 pub mod lookup_ip;
 // TODO: consider #[doc(hidden)]
