+++ tests/integration-tests/tests/client_future_tests.rs
@@ -132,7 +132,7 @@ fn test_query_tcp_ipv6() {
 #[cfg(feature = "dns-over-https-rustls")]
 fn test_query_https() {
     use rustls::{ClientConfig, OwnedTrustAnchor, RootCertStore};
-    use trust_dns_proto::https::HttpsClientStreamBuilder;
+    use trust_dns_proto::h2::HttpsClientStreamBuilder;
 
     const ALPN_H2: &[u8] = b"h2";
